requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP Requests
requests==2.31.0

# Environment Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3

# Date/Time utilities
python-dateutil==2.8.2
